lambda/processGeneration/requirements.txt
google-generativeai
pinecone
